% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_freq_factor.R
\name{ST_freq_factor}
\alias{ST_freq_factor}
\title{Frequency table for categorical data
Frequency table for categorical data.}
\usage{
ST_freq_factor(data, variable)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{tibble}.}

\item{variable}{Column (factor) in \code{data}.}
}
\value{
A \code{data.frame} with frequency, cumulative frequency, frequency percent and cumulative frequency
percent.
}
\description{
Frequency table for categorical data
Frequency table for categorical data.
}
\examples{
ST_freq_factor(iris,Species)
}
