% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_freq_numeric.R
\name{ST_freq_numeric}
\alias{ST_freq_numeric}
\title{Frequency table for continuous data
Frequency table for continuous data.}
\usage{
ST_freq_numeric(data, variable, bins = 5)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{tibble}.}

\item{variable}{Column in \code{data}.}

\item{bins}{Number of intervals into which the data must be split.}
}
\value{
A \code{data.frame} with frequency, cumulative frequency, frequency percent and cumulative frequency
percent.
}
\description{
Frequency table for continuous data
Frequency table for continuous data.
}
\examples{
ST_freq_numeric(mtcars, mpg,bins = 5)
}
